+++ device/topology/topology_discovery_wormhole.cpp
@@ -15,7 +15,7 @@
 
 #include "tracy.hpp"
 #include "umd/device/firmware/erisc_firmware.hpp"
-#include "umd/device/firmware/firmware_info_provider.hpp"
+#include "umd/device/firmware/firmware_info_provider_implementation.hpp"
 #include "umd/device/firmware/firmware_utils.hpp"
 #include "umd/device/soc_descriptor.hpp"
 #include "umd/device/topology/topology_discovery_error.hpp"
